feat: Many To One implementation

+++ Stock_Management/src/main/java/com/stock_management/dto/CountProductsDto.java
@@ -3,5 +3,5 @@ import lombok.Data;
 
 @Data
 public class CountProductsDto {
-    private int numberOfProducts;
+    private Long numberOfProducts;
 }
